moved viewsModel.py and models.py into /app from /models
