+++ app/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from flask_login import LoginManager
 from sqlalchemy.orm import sessionmaker
-from app.models import models
+from app import models
 #from flask_sqlalchemy import SQLAlchemy
 
 app = Flask(__name__)
@@ -16,7 +16,7 @@ Session.configure(bind=models.engine)
 sql_session = Session()
 
 from app import views
-from app.models import models #imports views.py from the app package
+#from models import models #imports views.py from the app package
