+++ app/views.py
@@ -2,13 +2,13 @@
 from flask import render_template, flash, session, redirect, request, Flask, url_for, jsonify
 from flask_oauthlib.client import OAuth
 from app import app, sql_session, login_manager, viewsClasses
-from .models import models, viewsModel
+from app import models, viewsModel
 from .myfunctions import sortnumbers
 from json import loads
 from sqlalchemy import update, func
-from urllib.request import Request, urlopen
-from urllib.parse import unquote
-from urllib.error import URLError
+#from urllib.request import Request, urlopen
+#from urllib.parse import unquote
+#from urllib.error import URLError
 import datetime, re
 
 GOOGLE_CLIENT_ID = '273956341734-jhk5ekhmrbeebqfef7d6f3vfeqf0aprg.apps.googleusercontent.com'
